# openapi.yaml
openapi: 3.0.3
info:
  title: RSO Events API
  version: 1.0.0
  description: MVP API for student squads events management.
servers:
  - url: http://localhost:8080
paths:
  /healthz:
    get:
      summary: Health check
      responses:
        "200":
          description: OK
  /api/v1/auth/register:
    post:
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, last_name, first_name]
              properties:
                email: { type: string, format: email }
                password: { type: string }
                last_name: { type: string }
                first_name: { type: string }
                middle_name: { type: string }
      responses:
        "201":
          description: Created
  /api/v1/auth/login:
    post:
      summary: Login by email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: OK
  /api/v1/auth/refresh:
    post:
      summary: Refresh access and refresh tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        "200":
          description: OK
  /api/v1/auth/logout:
    post:
      summary: Logout and revoke tokens
      security: [{ bearerAuth: [] }]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token: { type: string }
      responses:
        "200":
          description: OK
  /api/v1/events:
    get:
      summary: List events
      parameters:
        - in: query
          name: level
          schema: { type: string }
        - in: query
          name: type
          schema: { type: string }
        - in: query
          name: search
          schema: { type: string }
      responses:
        "200":
          description: OK
  /api/v1/me:
    get:
      summary: Current user profile
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
  /api/v1/portfolio:
    get:
      summary: Current user portfolio stats
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
  /api/v1/events/{eventID}/register:
    post:
      summary: Register current user for event
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: eventID
          required: true
          schema: { type: integer }
      responses:
        "201":
          description: Created
  /api/v1/attendance/scan:
    post:
      summary: Mark attendance by QR UUID
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [qr_code]
              properties:
                qr_code: { type: string, format: uuid }
      responses:
        "200":
          description: OK
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT